Wait for the time between timestamps when animating saved SVG.
Previously, the animation in a saved SVG would wait 500ms before
displaying the next frame, regardless of the amount of time elapsed
between the timestamps. This adjusts the logic to wait an amount of
time equal to the difference between timestamps, more accurately
reflecting the time as originally recorded.

Sub-second time differences are not captured by the parsed timestamps,
so multiple frames occuring in the same second will be displayed as fast
as possible. This could be improved by parsing out the Unix timestamp
from the URL instead of reading the text, but some URLs may not be
formatted correctly to allow this. This alternative is left as a future
enhancement.